Isolate preopen dashboard decision builders
The preopen dashboard service had accumulated artifact, QA, and execution-review construction alongside async data orchestration. This keeps the public service entrypoint stable while moving deterministic builders into focused modules so future UI/QA changes do not reopen the entire aggregation file.\n\nConstraint: Behavior must remain read-only and preserve app/routers/preopen.py response contracts.\nRejected: Split router/API contracts in the same change | unnecessary coupling and higher review risk.\nConfidence: high\nScope-risk: narrow\nReversibility: clean\nDirective: Keep broker/order/watch imports out of preopen dashboard builders; they are DTO-only and advisory-only.\nTested: uv run --group dev ruff check app/services/preopen_dashboard_service.py app/services/preopen_dashboard_artifacts.py app/services/preopen_dashboard_quality.py app/services/preopen_dashboard_execution_review.py; uv run --group test pytest tests/test_preopen_dashboard_service.py -q (26 passed)\nNot-tested: Full repository test suite
