+++ app/services/preopen_dashboard_service.py
@@ -6,7 +6,6 @@ Read-only. Never imports broker, order, watch, intent, or credential modules.
 from __future__ import annotations
 
 import logging
-from datetime import UTC, datetime
 from inspect import isawaitable
 from typing import Any
 
@@ -16,22 +15,13 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.models.news import NewsArticle
 from app.models.research_run import ResearchRun, ResearchRunCandidate
 from app.models.trading_decision import TradingDecisionProposal, TradingDecisionSession
-from app.schemas.execution_contracts import (
-    ExecutionGuard,
-    ExecutionReadiness,
-    OrderBasketPreview,
-    OrderPreviewLine,
-)
 from app.schemas.preopen import (
     CandidateSummary,
-    ExecutionReviewStage,
-    ExecutionReviewSummary,
     LinkedSessionRef,
     NewsArticlePreview,
     NewsReadinessSummary,
     NewsSourceCoverageSummary,
     PreopenArtifactReadinessItem,
-    PreopenArtifactSection,
     PreopenBriefingArtifact,
     PreopenBriefingRelevance,
     PreopenDecisionSessionCta,
@@ -39,9 +29,6 @@ from app.schemas.preopen import (
     PreopenMarketNewsBriefing,
     PreopenMarketNewsItem,
     PreopenMarketNewsSection,
-    PreopenQaCheck,
-    PreopenQaEvaluatorSummary,
-    PreopenQaScore,
     ReconciliationSummary,
 )
 from app.schemas.preopen_news_brief import KRPreopenNewsBrief
@@ -54,6 +41,9 @@ from app.services.market_news_briefing_formatter import (
     BriefingItem,
     format_market_news_briefing,
 )
+from app.services.preopen_dashboard_artifacts import _build_briefing_artifact
+from app.services.preopen_dashboard_execution_review import _build_execution_review
+from app.services.preopen_dashboard_quality import _build_qa_evaluator_summary
 from app.services.preopen_paper_approval_bridge import (
     build_preopen_paper_approval_bridge,
 )
@@ -429,791 +419,6 @@ def _build_news_brief(
     except Exception:
         logger.warning("Failed to build KR preopen news brief", exc_info=True)
         return None
-
-
-def _summarize_market_brief(market_brief: dict[str, Any] | None) -> str | None:
-    if not market_brief:
-        return None
-    summary = market_brief.get("summary")
-    if isinstance(summary, str) and summary.strip():
-        return summary
-    return None
-
-
-def _build_briefing_artifact(
-    *,
-    run: ResearchRun,
-    candidates: list[CandidateSummary],
-    reconciliations: list[ReconciliationSummary],
-    linked: list[LinkedSessionRef],
-    news: NewsReadinessSummary | None,
-    news_brief: KRPreopenNewsBrief | None,
-    market_news_briefing: PreopenMarketNewsBriefing | None,
-    source_warnings: list[str],
-    stage: str,
-) -> PreopenBriefingArtifact:
-    """Build the additive read-only preopen artifact from already-loaded data.
-
-    The artifact is a transport-level workflow summary only. It deliberately does
-    not query broker/order/watch providers or persist anything.
-    """
-
-    buy_candidates = [c for c in candidates if c.side == "buy"]
-    holding_actions = [c for c in candidates if c.side in {"sell", "none"}]
-    risk_notes = list(dict.fromkeys([*source_warnings]))
-    if news is None:
-        risk_notes.append("news_readiness_unavailable")
-    elif not news.is_ready:
-        risk_notes.extend(news.warnings or [f"news_{news.status}"])
-    if market_news_briefing is None:
-        risk_notes.append("market_news_briefing_unavailable")
-
-    news_status = news.status if news is not None else "unavailable"
-    readiness = [
-        PreopenArtifactReadinessItem(
-            key="research_run",
-            status="ready",
-            is_ready=True,
-            warnings=[],
-            details={"source_run_status": run.status},
-        ),
-        PreopenArtifactReadinessItem(
-            key="news",
-            status=news_status,  # type: ignore[arg-type]
-            is_ready=bool(news and news.is_ready),
-            warnings=list(news.warnings if news else ["news_readiness_unavailable"]),
-            details={
-                "latest_run_uuid": news.latest_run_uuid if news else None,
-                "latest_status": news.latest_status if news else None,
-                "source_counts": news.source_counts if news else {},
-                "source_coverage": [
-                    coverage.model_dump(mode="json")
-                    for coverage in (news.source_coverage if news else [])
-                ],
-            },
-        ),
-        PreopenArtifactReadinessItem(
-            key="cash",
-            status="unavailable",
-            is_ready=False,
-            warnings=["not_in_current_preopen_contract"],
-            details={},
-        ),
-        PreopenArtifactReadinessItem(
-            key="holdings",
-            status="partial" if holding_actions or reconciliations else "unavailable",
-            is_ready=bool(holding_actions or reconciliations),
-            warnings=[]
-            if holding_actions or reconciliations
-            else ["not_in_current_preopen_contract"],
-            details={
-                "holding_action_count": len(holding_actions),
-                "reconciliation_count": len(reconciliations),
-            },
-        ),
-        PreopenArtifactReadinessItem(
-            key="quotes",
-            status="unavailable",
-            is_ready=False,
-            warnings=["not_in_current_preopen_contract"],
-            details={},
-        ),
-    ]
-
-    market_news_count = (
-        sum(len(section.items) for section in market_news_briefing.sections)
-        if market_news_briefing
-        else 0
-    )
-    sections = [
-        PreopenArtifactSection(
-            section_id="market_news",
-            title="Market news briefing",
-            item_count=market_news_count,
-            status="ready" if market_news_count else "unavailable",
-            summary=(
-                f"{market_news_count} high-signal articles across "
-                f"{len(market_news_briefing.sections)} sections"
-                if market_news_briefing
-                else "Market news briefing is unavailable."
-            ),
-            items=[
-                {
-                    "section_id": section.section_id,
-                    "title": section.title,
-                    "item_count": len(section.items),
-                }
-                for section in (
-                    market_news_briefing.sections if market_news_briefing else []
-                )
-            ],
-        ),
-        PreopenArtifactSection(
-            section_id="new_buy_candidates",
-            title="New buy candidates",
-            item_count=len(buy_candidates),
-            status="ready" if buy_candidates else "unavailable",
-            summary=(
-                f"{len(buy_candidates)} buy candidates prepared before decision-session review."
-                if buy_candidates
-                else "No buy candidates in the current run."
-            ),
-            items=[
-                {
-                    "symbol": c.symbol,
-                    "confidence": c.confidence,
-                    "rationale": c.rationale,
-                    "proposed_price": str(c.proposed_price)
-                    if c.proposed_price is not None
-                    else None,
-                    "proposed_qty": str(c.proposed_qty)
-                    if c.proposed_qty is not None
-                    else None,
-                }
-                for c in buy_candidates[:5]
-            ],
-        ),
-        PreopenArtifactSection(
-            section_id="holdings_actions",
-            title="Holdings actions",
-            item_count=len(holding_actions) + len(reconciliations),
-            status="ready" if holding_actions or reconciliations else "unavailable",
-            summary=(
-                f"{len(holding_actions)} candidate actions and {len(reconciliations)} pending reconciliations."
-            ),
-            items=[
-                {"symbol": c.symbol, "side": c.side, "rationale": c.rationale}
-                for c in holding_actions[:5]
-            ]
-            + [
-                {
-                    "symbol": r.symbol,
-                    "classification": r.classification,
-                    "summary": r.summary,
-                }
-                for r in reconciliations[:5]
-            ],
-        ),
-    ]
-
-    cta = (
-        PreopenDecisionSessionCta(
-            state="linked_session_exists",
-            label="Open linked decision session",
-            run_uuid=run.run_uuid,
-            linked_session_uuid=linked[0].session_uuid,
-            requires_confirmation=False,
-        )
-        if linked
-        else PreopenDecisionSessionCta(
-            state="create_available",
-            label="Create decision session",
-            run_uuid=run.run_uuid,
-            requires_confirmation=True,
-        )
-    )
-
-    status = "ready"
-    if risk_notes or any(
-        not item.is_ready for item in readiness if item.key in {"research_run", "news"}
-    ):
-        status = "degraded"
-
-    news_summary = None
-    if news_brief is not None:
-        signal_count = len(news_brief.sector_flags) + len(news_brief.candidate_flags)
-        news_summary = (
-            f"News readiness is {news_brief.news_readiness}; "
-            f"{signal_count} advisory signals summarized."
-        )
-    elif news is not None:
-        news_summary = f"News readiness is {news.status}."
-
-    return PreopenBriefingArtifact(
-        status=status,  # type: ignore[arg-type]
-        run_uuid=run.run_uuid,
-        market_scope=run.market_scope,  # type: ignore[arg-type]
-        stage=stage,  # type: ignore[arg-type]
-        generated_at=run.generated_at,
-        source_run_status=run.status,
-        readiness=readiness,
-        market_summary=_summarize_market_brief(run.market_brief),
-        news_summary=news_summary,
-        sections=sections,
-        risk_notes=risk_notes,
-        cta=cta,
-        qa={
-            "read_only": True,
-            "mutation_paths": [],
-            "decision_session_created": False,
-            "source": "latest_open_research_run",
-        },
-    )
-
-
-def _qa_check(
-    check_id: str,
-    label: str,
-    status: str,
-    severity: str,
-    summary: str,
-    details: dict[str, Any] | None = None,
-) -> PreopenQaCheck:
-    return PreopenQaCheck(
-        id=check_id,
-        label=label,
-        status=status,  # type: ignore[arg-type]
-        severity=severity,  # type: ignore[arg-type]
-        summary=summary,
-        details=details,
-    )
-
-
-def _score_qa_checks(
-    checks: list[PreopenQaCheck],
-    *,
-    has_run: bool,
-) -> PreopenQaScore:
-    if not has_run:
-        return PreopenQaScore(
-            score=None,
-            grade="unavailable",
-            confidence="unavailable",
-            reason="no_open_preopen_run",
-        )
-    score = 100
-    for check in checks:
-        if check.status == "fail":
-            score -= 30
-        elif check.status == "warn":
-            score -= 10
-        elif check.status == "unknown":
-            score -= 5
-    score = max(0, min(100, score))
-    if score >= 90:
-        grade = "excellent"
-    elif score >= 75:
-        grade = "good"
-    elif score >= 50:
-        grade = "watch"
-    else:
-        grade = "poor"
-    confidence = "high" if score >= 75 else "medium" if score >= 50 else "low"
-    return PreopenQaScore(
-        score=score,
-        grade=grade,  # type: ignore[arg-type]
-        confidence=confidence,  # type: ignore[arg-type]
-        reason="deterministic checks over already-loaded preopen response data",
-    )
-
-
-def _build_qa_evaluator_summary(
-    *,
-    has_run: bool,
-    generated_at: datetime | None,
-    candidate_count: int,
-    reconciliation_count: int,
-    candidates: list[CandidateSummary],
-    reconciliations: list[ReconciliationSummary],
-    linked: list[LinkedSessionRef],
-    news: NewsReadinessSummary | None,
-    market_news_briefing: PreopenMarketNewsBriefing | None,
-    briefing_artifact: PreopenBriefingArtifact | None,
-    advisory_skipped_reason: str | None,
-) -> PreopenQaEvaluatorSummary:
-    """Build a deterministic, read-only QA summary from already-loaded DTOs only."""
-
-    coverage: dict[str, Any] = {
-        "candidate_count": candidate_count,
-        "candidate_items": len(candidates),
-        "reconciliation_count": reconciliation_count,
-        "reconciliation_items": len(reconciliations),
-        "linked_session_count": len(linked),
-        "news_status": news.status if news else "unavailable",
-        "market_news_sections": len(market_news_briefing.sections)
-        if market_news_briefing
-        else 0,
-        "briefing_artifact_status": briefing_artifact.status
-        if briefing_artifact
-        else "unavailable",
-        "advisory_only": True,
-        "execution_allowed": False,
-        "advisory_skipped_reason": advisory_skipped_reason,
-    }
-
-    if not has_run:
-        checks = [
-            _qa_check(
-                "has_open_run",
-                "Open preopen run",
-                "fail",
-                "high",
-                "No open preopen research run is available.",
-                {"reason": advisory_skipped_reason or "no_open_preopen_run"},
-            ),
-            _qa_check(
-                "actionability_guardrail",
-                "Actionability guardrail",
-                "pass",
-                "info",
-                "Evaluator is advisory-only and execution remains disabled.",
-                {"advisory_only": True, "execution_allowed": False},
-            ),
-        ]
-        return PreopenQaEvaluatorSummary(
-            status="unavailable",
-            generated_at=generated_at,
-            overall=_score_qa_checks(checks, has_run=False),
-            checks=checks,
-            blocking_reasons=["no_open_preopen_run"],
-            warnings=[],
-            coverage=coverage,
-        )
-
-    checks: list[PreopenQaCheck] = [
-        _qa_check(
-            "has_open_run",
-            "Open preopen run",
-            "pass",
-            "info",
-            "Open preopen research run loaded for read-only evaluation.",
-        )
-    ]
-
-    if briefing_artifact is None:
-        checks.append(
-            _qa_check(
-                "briefing_artifact_available",
-                "Briefing artifact",
-                "fail",
-                "high",
-                "Preopen briefing artifact is missing.",
-            )
-        )
-    elif briefing_artifact.status == "ready":
-        checks.append(
-            _qa_check(
-                "briefing_artifact_available",
-                "Briefing artifact",
-                "pass",
-                "info",
-                "Briefing artifact is ready.",
-                {"status": briefing_artifact.status},
-            )
-        )
-    elif briefing_artifact.status == "degraded":
-        checks.append(
-            _qa_check(
-                "briefing_artifact_available",
-                "Briefing artifact",
-                "warn",
-                "medium",
-                "Briefing artifact is degraded and should be reviewed.",
-                {"risk_notes": list(briefing_artifact.risk_notes)},
-            )
-        )
-    else:
-        checks.append(
-            _qa_check(
-                "briefing_artifact_available",
-                "Briefing artifact",
-                "fail",
-                "high",
-                f"Briefing artifact status is {briefing_artifact.status}.",
-            )
-        )
-
-    if news is None:
-        checks.append(
-            _qa_check(
-                "news_readiness",
-                "News readiness",
-                "fail",
-                "high",
-                "News readiness is unavailable.",
-            )
-        )
-    elif news.status == "ready" and news.is_ready:
-        checks.append(
-            _qa_check(
-                "news_readiness",
-                "News readiness",
-                "pass",
-                "info",
-                "News readiness is fresh.",
-                {"source_counts": news.source_counts},
-            )
-        )
-    elif news.status == "stale" or news.is_stale:
-        checks.append(
-            _qa_check(
-                "news_readiness",
-                "News readiness",
-                "fail",
-                "high",
-                "News readiness is stale; review before relying on recommendations.",
-                {"warnings": list(news.warnings)},
-            )
-        )
-    else:
-        checks.append(
-            _qa_check(
-                "news_readiness",
-                "News readiness",
-                "fail",
-                "high",
-                "News readiness is unavailable.",
-                {"warnings": list(news.warnings)},
-            )
-        )
-
-    if len(candidates) != candidate_count:
-        checks.append(
-            _qa_check(
-                "candidate_coverage",
-                "Candidate coverage",
-                "fail",
-                "high",
-                "Candidate count does not match candidate list length.",
-                {"candidate_count": candidate_count, "items": len(candidates)},
-            )
-        )
-    elif candidate_count == 0:
-        checks.append(
-            _qa_check(
-                "candidate_coverage",
-                "Candidate coverage",
-                "warn",
-                "medium",
-                "No candidates are present in the latest run.",
-            )
-        )
-    else:
-        checks.append(
-            _qa_check(
-                "candidate_coverage",
-                "Candidate coverage",
-                "pass",
-                "info",
-                f"{candidate_count} candidates are present.",
-            )
-        )
-
-    if len(reconciliations) != reconciliation_count:
-        checks.append(
-            _qa_check(
-                "reconciliation_coverage",
-                "Reconciliation coverage",
-                "fail",
-                "high",
-                "Reconciliation count does not match reconciliation list length.",
-                {
-                    "reconciliation_count": reconciliation_count,
-                    "items": len(reconciliations),
-                },
-            )
-        )
-    else:
-        checks.append(
-            _qa_check(
-                "reconciliation_coverage",
-                "Reconciliation coverage",
-                "pass" if reconciliation_count else "skipped",
-                "info",
-                f"{reconciliation_count} pending reconciliations summarized.",
-            )
-        )
-
-    if market_news_briefing is None:
-        checks.append(
-            _qa_check(
-                "market_news_briefing",
-                "Market news briefing",
-                "warn",
-                "medium",
-                "Market news briefing is unavailable; raw news readiness may still be shown.",
-            )
-        )
-    else:
-        market_news_count = sum(
-            len(section.items) for section in market_news_briefing.sections
-        )
-        checks.append(
-            _qa_check(
-                "market_news_briefing",
-                "Market news briefing",
-                "pass" if market_news_count else "warn",
-                "low" if market_news_count else "medium",
-                f"{market_news_count} market news briefing items summarized.",
-                {"sections": len(market_news_briefing.sections)},
-            )
-        )
-
-    checks.append(
-        _qa_check(
-            "linked_session_safety",
-            "Linked session safety",
-            "pass" if linked else "warn",
-            "info" if linked else "low",
-            "Linked sessions are read-only summaries; evaluator does not create sessions."
-            if linked
-            else "No linked decision session found; evaluator did not create one.",
-            {"linked_session_count": len(linked)},
-        )
-    )
-    checks.append(
-        _qa_check(
-            "actionability_guardrail",
-            "Actionability guardrail",
-            "pass",
-            "info",
-            "QA evaluator is advisory-only and execution remains disabled.",
-            {"advisory_only": True, "execution_allowed": False},
-        )
-    )
-
-    overall = _score_qa_checks(checks, has_run=True)
-    blocking_reasons = [check.id for check in checks if check.status == "fail"]
-    warnings = [
-        check.summary for check in checks if check.status in {"warn", "unknown"}
-    ]
-    status = (
-        "ready"
-        if not blocking_reasons and overall.score is not None and overall.score >= 75
-        else "needs_review"
-    )
-    return PreopenQaEvaluatorSummary(
-        status=status,
-        generated_at=generated_at or datetime.now(UTC),
-        overall=overall,
-        checks=checks,
-        blocking_reasons=blocking_reasons,
-        warnings=warnings,
-        coverage=coverage,
-    )
-
-
-def _build_execution_review(
-    *,
-    has_run: bool,
-    market_scope: str | None,
-    stage: str | None,
-    candidates: list[CandidateSummary],
-    reconciliations: list[ReconciliationSummary],
-    news: NewsReadinessSummary | None,
-    briefing_artifact: PreopenBriefingArtifact | None,
-) -> ExecutionReviewSummary:
-    """Deterministic, read-only execution review built from already-loaded data.
-
-    Always returns an ``advisory_only=True`` summary with conservative
-    ``ExecutionGuard`` defaults. No broker / order / watch / credential code is
-    consulted: ``cash``, ``holdings``, ``quotes``, ``broker_order``, and
-    ``watch`` are reported as ``unavailable`` with the
-    ``not_in_current_preopen_contract`` warning so the UI can render them as
-    degraded rather than silently dropping them.
-    """
-
-    blocking_reasons: list[str] = ["mvp_read_only"]
-    warnings: list[str] = []
-
-    if not has_run:
-        blocking_reasons.append("no_open_preopen_run")
-
-    news_status = news.status if news is not None else "unavailable"
-    if news is None or not news.is_ready:
-        blocking_reasons.append(f"news_{news_status}")
-
-    buy_candidates = [c for c in candidates if c.side == "buy"]
-
-    stages: list[ExecutionReviewStage] = []
-
-    # 1) data / news
-    if news is None:
-        data_status: str = "unavailable"
-        data_summary = "News readiness is unavailable."
-        data_warnings = ["news_readiness_unavailable"]
-    elif news.is_ready:
-        data_status = "ready"
-        data_summary = "News readiness is fresh."
-        data_warnings = []
-    else:
-        data_status = "degraded"
-        data_summary = f"News readiness is {news.status}."
-        data_warnings = list(news.warnings)
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="data_news",
-            label="Data / news readiness",
-            status=data_status,  # type: ignore[arg-type]
-            summary=data_summary,
-            warnings=data_warnings,
-            details={"news_status": news_status},
-        )
-    )
-
-    # 2) candidate review
-    if not has_run:
-        candidate_status = "unavailable"
-        candidate_summary = "No open preopen research run."
-    elif not candidates:
-        candidate_status = "degraded"
-        candidate_summary = "Open run has no candidates."
-    else:
-        candidate_status = "ready"
-        candidate_summary = f"{len(candidates)} candidates ({len(buy_candidates)} buy)."
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="candidate_review",
-            label="Candidate review",
-            status=candidate_status,  # type: ignore[arg-type]
-            summary=candidate_summary,
-            details={
-                "candidate_count": len(candidates),
-                "buy_candidate_count": len(buy_candidates),
-            },
-        )
-    )
-
-    # 3) cash / holdings / quotes (always unavailable in MVP read-only)
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="cash_holdings_quotes",
-            label="Cash / holdings / quotes check",
-            status="unavailable",
-            summary="Live cash, holdings, and quotes lookups are not wired in this MVP.",
-            warnings=["not_in_current_preopen_contract"],
-            details={},
-        )
-    )
-
-    # 4) basket preview
-    basket_preview: OrderBasketPreview | None = None
-    if buy_candidates:
-        readiness_for_basket = ExecutionReadiness(
-            account_mode="db_simulated",
-            execution_source="preopen",
-            is_ready=False,
-            guard=ExecutionGuard(
-                execution_allowed=False,
-                approval_required=True,
-                blocking_reasons=["mvp_read_only"],
-            ),
-        )
-        basket_preview = OrderBasketPreview(
-            account_mode="db_simulated",
-            execution_source="preopen",
-            readiness=readiness_for_basket,
-            lines=[
-                OrderPreviewLine(
-                    symbol=c.symbol,
-                    market=market_scope or "kr",
-                    side="buy",
-                    account_mode="db_simulated",
-                    execution_source="preopen",
-                    quantity=c.proposed_qty,
-                    limit_price=c.proposed_price,
-                    currency=c.currency,
-                    guard=ExecutionGuard(
-                        execution_allowed=False,
-                        approval_required=True,
-                        blocking_reasons=["mvp_read_only"],
-                    ),
-                    rationale=[c.rationale] if c.rationale else [],
-                )
-                for c in buy_candidates
-            ],
-            basket_warnings=["mvp_read_only"],
-        )
-        basket_status = "ready"
-        basket_summary = (
-            f"{len(buy_candidates)} buy candidates rendered as a basket preview."
-        )
-    elif has_run:
-        basket_status = "degraded"
-        basket_summary = "No buy candidates available for basket preview."
-    else:
-        basket_status = "unavailable"
-        basket_summary = "No open run to derive a basket from."
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="basket_preview",
-            label="Basket preview",
-            status=basket_status,  # type: ignore[arg-type]
-            summary=basket_summary,
-            details={"line_count": len(basket_preview.lines) if basket_preview else 0},
-        )
-    )
-
-    # 5) approval required
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="approval_required",
-            label="Approval required",
-            status="pending",
-            summary=(
-                "Mock execution requires later explicit operator approval. "
-                "This page does not submit orders."
-            ),
-            details={"advisory_only": True, "execution_allowed": False},
-        )
-    )
-
-    # 6) post-order reconcile
-    pending_recs = len(reconciliations)
-    if not has_run:
-        recon_status = "unavailable"
-        recon_summary = "No open run to drive reconciliation."
-    elif pending_recs == 0:
-        recon_status = "skipped"
-        recon_summary = "No pending reconciliations on the latest run."
-    else:
-        recon_status = "pending"
-        recon_summary = f"{pending_recs} pending reconciliations to review."
-    stages.append(
-        ExecutionReviewStage(
-            stage_id="post_order_reconcile",
-            label="Post-order reconciliation",
-            status=recon_status,  # type: ignore[arg-type]
-            summary=recon_summary,
-            details={"pending_reconciliation_count": pending_recs},
-        )
-    )
-
-    # Lift any per-stage warnings into the summary warnings (de-duplicated).
-    for s in stages:
-        for w in s.warnings:
-            if w not in warnings:
-                warnings.append(w)
-
-    readiness = ExecutionReadiness(
-        account_mode="db_simulated",
-        execution_source="preopen",
-        is_ready=False,
-        guard=ExecutionGuard(
-            execution_allowed=False,
-            approval_required=True,
-            blocking_reasons=blocking_reasons,
-            warnings=warnings,
-        ),
-        notes=[
-            "Advisory read-only review; no broker submit on this page.",
-        ],
-    )
-
-    notes: list[str] = [
-        "advisory_only",
-        "no_live_execution",
-        "mock_execution_requires_explicit_approval",
-    ]
-
-    return ExecutionReviewSummary(
-        readiness=readiness,
-        stages=stages,
-        basket_preview=basket_preview,
-        blocking_reasons=blocking_reasons,
-        warnings=warnings,
-        notes=notes,
-    )
 
 
 async def get_latest_preopen_dashboard(
